chore: feedback from peer-review.
logion-network/logion-internal#717
